Update to new minimax APIs, and simplify noisy_moves.
The beetle cover procrastination is an obvious horizon effect problem of the current engine. Using this simple noisy move heuristic seems to create an even larger horizon effect of procrastinating attacking the enemy queen. Still beats Mzinga.

+++ src/uhp_server.rs
@@ -96,7 +96,7 @@ impl UhpServer {
 
     // Bonus undocumented command.
     fn perft(&mut self, args: &str) -> Result<()> {
-        let depth = args.parse::<usize>().unwrap_or(20);
+        let depth = args.parse::<u8>().unwrap_or(20);
         let mut b = self.board.as_ref().ok_or(UhpError::GameNotStarted)?.clone();
         minimax::perft::<Rules>(&mut b, depth, false);
         Ok(())
